Fixing if..elseif()
git-svn-id: svn://svn.code.sf.net/p/pebl/code-0/trunk@146 12333153-ee01-4b9b-8cc9-3b69a7b36d39

+++ src/base/lex.yy.c
@@ -1310,7 +1310,7 @@ static int yy_get_next_buffer (void)
 
 	else
 		{
-			int num_to_read =
+			size_t num_to_read =
 			YY_CURRENT_BUFFER_LVALUE->yy_buf_size - number_to_move - 1;
 
 		while ( num_to_read <= 0 )
